multi-tenant-oauth/requirements.txt
# Multi-Tenant OAuth Social Media Automation
# Python 3.10+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Redis (caching & background jobs)
redis>=5.0.1
hiredis>=2.3.2

# Background Jobs
celery>=5.3.6
flower>=2.0.1  # Celery monitoring

# Security & Encryption
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# OAuth & HTTP
httpx>=0.26.0
authlib>=1.3.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# OpenAI (existing)
openai>=1.12.0

# Image Processing (existing)
pillow>=10.2.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Monitoring & Logging
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx[http2]>=0.26.0  # For testing

# Development
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0
